+++ sablecc/grammar/postfix/Compiler.java
@@ -3,8 +3,8 @@ package postfix;
 import postfix.semantics.SymbolTable;
 import postfix.semantics.Exceptions.InvalidExpressionException;
 import postfix.parser.*;
-import postfix.semantics.visitors.CodeGen;
 import postfix.semantics.visitors.SemanticVisitor;
+import postfix.semantics.visitors.CodeGen.CodeGen;
 import postfix.lexer.*;
 import postfix.node.*;
 import java.io.*;
