+++ sablecc/grammar/postfix/semantics/visitors/CodeGen/CommonCodeGen.java
@@ -1,0 +1,19 @@
+package postfix.semantics.visitors.CodeGen;
+
+import postfix.analysis.DepthFirstAdapter;
+import postfix.semantics.SymbolTable;
+
+public class CommonCodeGen extends DepthFirstAdapter {
+    // Variables
+    SymbolTable symbolTable;
+
+    String program;
+    
+    //Constructor
+    public CommonCodeGen() {};
+
+    public CommonCodeGen(SymbolTable symbolTable, String program) {
+        this.symbolTable = symbolTable;
+        this.program = program;
+    }
+}
