+++ sablecc/grammar/postfix/semantics/visitors/CodeGen/MethodDclCodeGen.java
@@ -1,0 +1,5 @@
+package postfix.semantics.visitors.CodeGen;
+
+public class MethodDclCodeGen extends CommonCodeGen {
+    
+}
